+++ README.md
@@ -1,0 +1,3 @@
+# SuperFrog Scheduler
+
+This is a single-page application that allows customers to request a SuperFrog to appear at their event, SuperFrog Students to sign up for an approved appearance, and the Spirit Director to manage submitted requests and users of this system.
