+++ src/components/Tables/CustomerTable.vue
